Small fixes and XMPP Transport
* Events are sent as objects internally, transports should take care of
serialization and deserialization
* Event names are processed as lists internally to increase efficiency,
they should be transformed to a string before sending (up to the transport).
* ForSubscription and similar fields will be processed as strings by now
because it makes them easier to handle. Besides, it's almost impossible
to find a standard way to convert **ALL** the event names inside the
data field. In an xml-based protocol like XMPP we could ensure event
names are always converted, but in JSON we can't.